added voting functionality

+++ src/main/java/com/twinkles/simpoprojectjava/model/Candidate.java
@@ -12,5 +12,5 @@ public class Candidate {
     private String fullName;
     private Party party;
     private VoteCategory voteCategory;
-    private int voteCount;
+    private long voteCount;
 }
